+++ src/init.c
@@ -2,9 +2,13 @@
 #include <Rinternals.h>
 
 #include "keypress.h"
+#include "keypress-internal.h"
 
 static const R_CallMethodDef callMethods[]  = {
   {"keypress", (DL_FUNC) &keypress, 1},
+  {"save_term_status", (DL_FUNC) &save_term_status, 0},
+  {"restore_term_status", (DL_FUNC) &restore_term_status, 0},
+  {"set_term_echo", (DL_FUNC) &set_term_echo, 1},
   {NULL, NULL, 0}
 };