+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # development version
 
+- Allow turning echo off for non-blocking Linux and macOS terminals
+  (#11, @weshinsley)
 - Support non-blocking press on Windows (Rterm)
 
 # 1.2.0